% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{atc_cache_dir}
\alias{atc_cache_dir}
\title{Get or create cache directory}
\usage{
atc_cache_dir()
}
\value{
Character; path to cache directory
}
\description{
Get or create cache directory
}
\keyword{internal}
