% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{atc_user_agent}
\alias{atc_user_agent}
\title{Get user agent string for HTTP requests}
\usage{
atc_user_agent()
}
\value{
Character; user agent identification
}
\description{
Get user agent string for HTTP requests
}
\keyword{internal}
