% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{atc_write_csv}
\alias{atc_write_csv}
\title{Write codes and DDD tables to CSV files}
\usage{
atc_write_csv(x, dir = ".", stamp = TRUE)
}
\arguments{
\item{x}{list with 'codes' and 'ddd'}

\item{dir}{output directory}

\item{stamp}{logical date-stamped filenames}
}
\value{
character vector of file paths (invisible)
}
\description{
Write codes and DDD tables to CSV files
}
