% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-package.R
\docType{package}
\name{atcddd}
\alias{atcddd-package}
\alias{atcddd}
\title{atcddd: WHO ATC/DDD Crawler and Parser}
\description{
Crawl and parse the WHO ATC/DDD index (whocc.no) into tidy tables with
robust HTTP (retries, timeouts, UA), on-disk caching, rate limiting,
iterative traversal, and resilient parsing. Includes checksum manifests.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://vanhungtran.github.io/atcddd}
  \item Report bugs at \url{https://github.com/vanhungtran/atcddd/issues}
}

}
\author{
\strong{Maintainer}: Lucas VHH TRAN \email{tranhungydhcm@gmail.com}

}
