% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{.empty_atc_from_text_result}
\alias{.empty_atc_from_text_result}
\title{Zero-row tibble template for atc_from_text() return value}
\usage{
.empty_atc_from_text_result()
}
\description{
Zero-row tibble template for atc_from_text() return value
}
\keyword{internal}
