% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{atc_level}
\alias{atc_level}
\title{Determine ATC level from code}
\usage{
atc_level(code)
}
\arguments{
\item{code}{character vector of ATC codes}
}
\value{
integer vector with values 1..5 or NA for invalid/unknown
}
\description{
Determine ATC level from code
}
\keyword{internal}
