% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{atc_parent}
\alias{atc_parent}
\title{Get parent ATC code}
\usage{
atc_parent(code)
}
\arguments{
\item{code}{character vector of ATC codes}
}
\value{
character vector of parent codes or NA for Level 1/invalid
}
\description{
Get parent ATC code
}
\keyword{internal}
