% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{is_valid_atc_code}
\alias{is_valid_atc_code}
\alias{is_valid_atc}
\title{Validate ATC Code Format}
\usage{
is_valid_atc_code(x)

is_valid_atc(x)
}
\arguments{
\item{x}{Character; a potential ATC code to validate}
}
\value{
Logical; TRUE if valid format, FALSE otherwise
}
\description{
Checks if a string is a valid ATC code format (uppercase alphanumeric).
ATC codes can be 1-7 characters long, following the WHO hierarchy:
\itemize{
\item Level 1: 1 character (e.g., "N")
\item Level 2: 3 characters (e.g., "N02")
\item Level 3: 4 characters (e.g., "N02B")
\item Level 4: 5 characters (e.g., "N02BE")
\item Level 5: 7 characters (e.g., "N02BE01")
}
}
\examples{
is_valid_atc_code("N02BE01")  # TRUE - valid Level 5 code
is_valid_atc_code("C")        # TRUE - valid Level 1 code
is_valid_atc_code("n02be01")  # FALSE - must be uppercase
is_valid_atc_code("N02-BE01") # FALSE - no special characters

}
