% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{is_valid_atc_code}
\alias{is_valid_atc_code}
\alias{is_valid_atc}
\title{Validate ATC Code Format}
\usage{
is_valid_atc_code(x)

is_valid_atc(x)
}
\arguments{
\item{x}{Character; a potential ATC code to validate (scalar)}
}
\value{
Logical; TRUE if valid format, FALSE otherwise
}
\description{
Checks if a string is a valid WHO ATC code by enforcing
the official level-specific patterns (uppercase only):
\itemize{
\item Level 1: \verb{^[A-Z]$}
\item Level 2: \verb{^[A-Z][0-9]\{2\}$}
\item Level 3: \verb{^[A-Z][0-9]\{2\}[A-Z]$}
\item Level 4: \verb{^[A-Z][0-9]\{2\}[A-Z]\{2\}$}
\item Level 5: \verb{^[A-Z][0-9]\{2\}[A-Z]\{2\}[0-9]\{2\}$}
}
}
\examples{
is_valid_atc_code("N02BE01")  # TRUE - valid Level 5 code
is_valid_atc_code("C")        # TRUE - valid Level 1 code
is_valid_atc_code("n02be01")  # FALSE - must be uppercase
is_valid_atc_code("N02-BE01") # FALSE - no special characters

}
